% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_returns}
\alias{calculate_returns}
\title{Takes a vector and calculates return}
\usage{
calculate_returns(prices, type = "simple")
}
\arguments{
\item{prices}{\itemize{
\item Vector with prices
}}

\item{type}{\itemize{
\item Type of return (simple / log)
}}
}
\value{
Vector of returns
}
\description{
Just a simple function for calculating returns.
}
