% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_if_date}
\alias{check_if_date}
\title{Checks if supplied object is a date object}
\usage{
check_if_date(date)
}
\arguments{
\item{date}{\itemize{
\item Object to check if is date
}}
}
\value{
a date object.
}
\description{
Tries to parse is as a date if not
}
