% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_JSON}
\alias{date_to_JSON}
\title{Converts date to JSON formate}
\usage{
date_to_JSON(date)
}
\arguments{
\item{date}{\itemize{
\item The date to convert
}}
}
\value{
date converted to JSON in string format
}
\description{
Converts date to JSON formate
}
