% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{debug_msg}
\alias{debug_msg}
\title{Printing a debug message}
\usage{
debug_msg(test, msg)
}
\arguments{
\item{test}{\itemize{
\item Boolean if the message should be printed
}}

\item{msg}{\itemize{
\item String with message
}}
}
\description{
Printing a debug message
}
