% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{dg_fetch_headers}
\alias{dg_fetch_headers}
\title{Fetches the headers of a json object}
\usage{
dg_fetch_headers(json_like)
}
\arguments{
\item{json_like}{\itemize{
\item json_like object, in reality a nested list
}}
}
\value{
list of headers
}
\description{
Fetches the headers of a json object
}
