% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{dg_to_dataframe}
\alias{dg_to_dataframe}
\title{Converts a json object to a dataframe}
\usage{
dg_to_dataframe(json_like)
}
\arguments{
\item{json_like}{\itemize{
\item json_like object, in reality a nested list
}}
}
\value{
dataframe
}
\description{
Converts a json object to a dataframe
}
