% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{done_msg}
\alias{done_msg}
\title{Printing a msg when something is done}
\usage{
done_msg(msg)
}
\arguments{
\item{msg}{\itemize{
\item String with message
}}
}
\description{
Printing a msg when something is done
}
