% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_connection.R
\name{ek_set_APIKEY}
\alias{ek_set_APIKEY}
\title{Function to set api_key}
\usage{
ek_set_APIKEY(api_key = NULL, debug = FALSE)
}
\arguments{
\item{api_key}{\itemize{
\item The api_key to be used
}}

\item{debug}{\itemize{
\item If TRUE, prints debug messages
}}
}
\description{
This function is used to set the app_key to be used in the package. To get an app_key you need to have the Eikon
desktop app or the newer Refinitiv terminal installed. You can then get an app_key from the app by typing in
appkey in the search field and then creating a new app_key.
}
