% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_connection.R
\name{ek_set_port}
\alias{ek_set_port}
\title{Function to set api_port}
\usage{
ek_set_port(port = NULL)
}
\arguments{
\item{port}{\itemize{
\item The new port to use, (default NULL)
}}
}
\description{
Just needed to incase it doesn't work on the automatically detected port,
this can sometimes be the case after the system hibernates with Eikon or
the Refintiv terminal running.
}
