% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{fetch_timeseries}
\alias{fetch_timeseries}
\title{Fetches a timeseries given a CUSIP}
\usage{
fetch_timeseries(CUSIP, start_date, end_date)
}
\arguments{
\item{CUSIP}{\itemize{
\item Vector of CUSIPs
}}

\item{start_date}{\itemize{
\item The start date in date format to start the query
}}

\item{end_date}{\itemize{
\item The end date in date format to end the query
}}
}
\value{
Dataframe of the requested information
}
\description{
Combines get_datagrid and get_timeseries into one function that firstly uses the CUSIP to fetch RICS and then uses
the RICS to lookup the timeseries data between the given dates
}
