% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{get_datagrid}
\alias{get_datagrid}
\title{Fetch datagrid information from the Eikon API}
\usage{
get_datagrid(instrument, fields, ..., settings = list(raw = FALSE))
}
\arguments{
\item{instrument}{\itemize{
\item Vector of Char, can be CUSIP, PERMID, rics any identifer that the Eikon can handle
}}

\item{fields}{\itemize{
\item Vector of Char, fields to request from the datagrid
}}

\item{...}{\itemize{
\item List of named parameters, could be 'SDate' = '2021-07-01', 'EDate' = '2021-09-28', 'Frq' = 'D' for
daily data (Frq) with a given start (SDate) and end date (EDate). If no EDate is supplied, the function will
use todays date. You can pass other arguments like for instance curn = 'USD' to get the data in USD, change out USD
to any other currency to get the fields in that currency.
}}

\item{settings}{\itemize{
\item List of bool settings, possibilities list(raw = false, field_name = false):
raw : If the function should return the raw json (default false)
field_name : if the function should return the field names (default false)
}}
}
\value{
dataframe or a list of raw data. At the moment i do not parse any column to a specific type, so all
columns are of type character. This is something I might change in the future, but only if i find a robust way
of doing this.
}
\description{
Before this function words you need to run the function ek_app_key() with a working app_key from the Eikon/Refinitiv
desktop
}
\details{
This package downloads information from the Eikon datagrid. The function is a wrapper around the Rust
function that does the actual work. To use this function you simply need to have an Eikon APP key, which you
get from the Eikon desktop app or the newer Refinitiv terminal. The function will then fetch the information
from the Eikon datagrid and return it as a dataframe. The desktop app needs to be running for this to work.

Error messages are passed from Rust (unless of course the issue is R type releated, then it is catched
directly in this function), if the error message displayed is along the line of "... panicked" this means you have
gotten an error I did not expect at all and you should file an issue with reproducible code and the error message
at github.com/OliverFrisvoll/EikonDownloader/issues
}
