% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get timeseries data from Eikon using RIC as identifier}
\usage{
get_timeseries(
  rics,
  fields = "*",
  startdate,
  enddate = NULL,
  interval = "daily"
)
}
\arguments{
\item{rics}{\itemize{
\item Char vector of rics for the information requested
}}

\item{fields}{\itemize{
\item Fields to return, by default all fields are returned these field are different from the datagrid
fields
}}

\item{startdate}{\itemize{
\item Date, start date of the query as a date object (required)
}}

\item{enddate}{\itemize{
\item Date, end date of the query as a date object (optional, if not supplied, todays date is used)
}}

\item{interval}{\itemize{
\item char, interval of data: (minute / hour / daily / weekly / monthly / quarterly / yearly) not all
data is available for all intervals. For instance minute and hour data is only available one year back
}}
}
\value{
A dataframe with the data requested
}
\description{
Returns a timeseries of data for the given rics for the given timeperiod and interval.
The fields can be specified, by default it returns all the fields
}
