% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_request.R
\name{health_searchView}
\alias{health_searchView}
\title{Quick check if the application can be run.
TODO: Write documentation}
\usage{
health_searchView()
}
\description{
Quick check if the application can be run.
TODO: Write documentation
}
