% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_msg}
\alias{info_msg}
\title{Printing a msg to inform about something}
\usage{
info_msg(msg)
}
\arguments{
\item{msg}{\itemize{
\item String with message
}}
}
\description{
Printing a msg to inform about something
}
