% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_request.R
\name{json_builder}
\alias{json_builder}
\title{Builds a json like list for the Eikon Data API}
\usage{
json_builder(directions, payload)
}
\arguments{
\item{directions}{\itemize{
\item Where
}}

\item{payload}{\itemize{
\item What
}}
}
\value{
Nested list that fit into the JSON scheme
}
\description{
Builds a json like list for the Eikon Data API
}
