% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_request.R
\name{POST_searchView}
\alias{POST_searchView}
\title{POST request for searchview}
\usage{
POST_searchView(json, service = "/", debug = FALSE)
}
\arguments{
\item{json}{\itemize{
\item The JSON to send
}}

\item{service}{\itemize{
\item Which service to use
}}

\item{debug}{\itemize{
\item Should debugging be on or off
}}
}
\value{
returns the results
}
\description{
TODO: Write documentation
}
