% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{rows_pr_inst}
\alias{rows_pr_inst}
\title{Figures out how many rows pr instrument.}
\usage{
rows_pr_inst(kwargs, debug = FALSE)
}
\arguments{
\item{kwargs}{\itemize{
\item Keyword arguments that can be passed to results
}}

\item{debug}{\itemize{
\item debug or not, default FALSE
}}
}
\value{
\itemize{
\item Rows pr instrument
}
}
\description{
Figures out how many rows pr instrument.
}
