% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_view}
\alias{search_view}
\title{A function to search through Eikon !!WIP!!
TODO: Write documentation}
\usage{
search_view(..., API = "/", debug = FALSE)
}
\description{
parameters: Query, Filter, View, OrderBy, Boost, Select, Top, Skip, GroupBy, GroupCount
}
