% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_json_request.R
\name{send_json_request}
\alias{send_json_request}
\title{Send JSON request to the database}
\usage{
send_json_request(directions, payload, debug = FALSE)
}
\arguments{
\item{directions}{\itemize{
\item The location to send it
}}

\item{payload}{\itemize{
\item What you are looking for formatted in a specific way
}}

\item{debug}{\itemize{
\item Default to FALSE, turns on debugging messages.
}}
}
\value{
Returns the results from the query
}
\description{
Sends a post request to the API with the given payload and a direction
}
