% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_of_dates}
\alias{seq_of_dates}
\title{Divides a date interval into many intervals}
\usage{
seq_of_dates(start, end, interval, RFE)
}
\arguments{
\item{start}{\itemize{
\item date object start date
}}

\item{end}{\itemize{
\item date object end date
}}

\item{interval}{\itemize{
\item interval between dates
}}

\item{RFE}{\itemize{
\item How many rows should be allowed for each RIC
}}
}
\value{
a list of start and end dates in JSON like format
}
\description{
Divides a date interval into many intervals
}
