% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Converts a listed JSON like object to a dataframe}
\usage{
to_dataframe(snippet)
}
\arguments{
\item{snippet}{\itemize{
\item A snippet of the JSON
}}
}
\value{
data.frame
}
\description{
Converts a listed JSON like object to a dataframe
}
