% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{ts_payload_loop}
\alias{ts_payload_loop}
\title{Timeseries payload loop}
\usage{
ts_payload_loop(
  directions,
  rics,
  fields,
  interval,
  start,
  end,
  app_key,
  url,
  debug = FALSE
)
}
\arguments{
\item{directions}{\itemize{
\item Where to send the request
}}

\item{rics}{\itemize{
\item Rics of the companies that will be looked up
}}

\item{fields}{\itemize{
\item The fields to return for those companies
}}

\item{interval}{\itemize{
\item the interval
}}

\item{start}{\itemize{
\item Start date, JSON format
}}

\item{end}{\itemize{
\item End date, JSON format
}}

\item{app_key}{\itemize{
\item API key
}}

\item{url}{\itemize{
\item URL to send the request to
}}

\item{debug}{\itemize{
\item Debug mode, prints the payload and the results
}}
}
\value{
JSON results
}
\description{
Timeseries payload loop
}
