% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_connection.R
\name{ek_set_APIKEY}
\alias{ek_set_APIKEY}
\title{Function to set api_key}
\usage{
ek_set_APIKEY(api_key)
}
\arguments{
\item{api_key}{\itemize{
\item The api_key to be used
}}
}
\description{
Function to set api_key
}
