% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API_connection.R
\name{ek_set_port}
\alias{ek_set_port}
\title{Function to set api_port}
\usage{
ek_set_port(port = NULL)
}
\arguments{
\item{port}{\itemize{
\item The new port to use, (default NULL)
}}
}
\description{
Only needed if the Refinitiv Terminal is running on a port that
is not between 9000L and 9011L
}
