% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{get_datagrid}
\alias{get_datagrid}
\title{Fetching information from the datagrid}
\usage{
get_datagrid(instrument, fields, ..., settings = list(raw = FALSE))
}
\arguments{
\item{instrument}{\itemize{
\item Vector of Char, can be CUSIP, rics
}}

\item{fields}{\itemize{
\item Vector of Char, fields to request from the datagrid
}}

\item{...}{\itemize{
\item List of named parameters, could be 'SDate' = '2021-07-01', 'EDate' = '2021-09-28', 'Frq' = 'D' for
daily data (Frq) with a given start (SDate) and end date (EDate)
}}

\item{settings}{\itemize{
\item List of bool settings, possibilities:
raw : If the function should return the raw json (default false)
field_name : if the function should return the field names (default false)
}}
}
\value{
dataframe of the information requested
}
\description{
Tries to fetch information from the Eikon datagrid location. The direction that seems to work for this is
'DataGrid_StandardAsync' even though I haven't implimented Async possibilities, could possibly do it with the
packages promises, but I doesn't seem necessary at the moment. If needed it should be implimented into a loop
for the send_json_request function.
}
