% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{get_datagrid}
\alias{get_datagrid}
\title{Fetching information from the datagrid}
\usage{
get_datagrid(instrument, fields)
}
\arguments{
\item{instrument}{\itemize{
\item Vector of Char, can be CUSIP, rics
}}

\item{fields}{\itemize{
\item Vector of Char, fields to request from the datagrid
}}
}
\value{
dataframe of the information requested
}
\description{
Tries to fetch information from the Eikon datagrid location. The direction that seems to work for this is
'DataGrid_StandardAsync' even though I haven't implimented Async possibilities, could possibly do it with the
packages promises, but I doesn't seem necessary at the moment. If needed it should be implimented into a loop
for the send_json_request function.
}
