% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get timeseries data from Eikon using RIC as identifier}
\usage{
get_timeseries(rics, fields = "*", startdate, enddate, interval = "daily")
}
\arguments{
\item{rics}{\itemize{
\item Char vector of rics for the information requested
}}

\item{fields}{\itemize{
\item Fields to return, can be found on Refinitiv Workspace
}}

\item{startdate}{\itemize{
\item Date, start date of the query
}}

\item{enddate}{\itemize{
\item Date, end date of the query
}}

\item{interval}{\itemize{
\item char, interval of data: (minute / hour / daily / weekly / monthly / quarterly / yearly)
}}
}
\value{
A dataframe with the data requested
}
\description{
Returns a timeseries of data for the given rics for the given timeperiod and interval.
The fields can be specified, by default it returns all the fields
}
