% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Fetch timeseries of data}
\usage{
get_timeseries(rics, fields = "*", startdate, enddate, interval = "daily")
}
\arguments{
\item{rics}{\itemize{
\item Vector of rics for the information requested
}}

\item{fields}{\itemize{
\item Fields to return
}}

\item{startdate}{\itemize{
\item Start date of the query
}}

\item{enddate}{\itemize{
\item End date of the query
}}

\item{interval}{\itemize{
\item char, interval of data: (minute / hour / daily / weekly / monthly / quarterly / yearly)
}}
}
\value{
A dataframe with the data requested
}
\description{
Returns a timeseries of data based upon the variables given
}
