% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_request.R
\name{Health_searchView}
\alias{Health_searchView}
\title{Quick check if the application can be run.
TODO: Write documentation}
\usage{
Health_searchView()
}
\description{
Quick check if the application can be run.
TODO: Write documentation
}
