% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_view}
\alias{search_view}
\title{A function to search through Eikon !!WIP!!
TODO: Write documentation}
\usage{
search_view(API = "/", debug = FALSE, ...)
}
\arguments{
\item{API}{\itemize{
\item API to use
}}

\item{debug}{\itemize{
\item If debugging should be turned on.
}}

\item{...}{\itemize{
\item Query, Filter, View, OrderBy, Boost, Select, Top, Skip, GroupBy, GroupCount
}}
}
\value{
The result of the query
}
\description{
A function to search through Eikon !!WIP!!
TODO: Write documentation
}
