% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_request.R
\name{send_json_request}
\alias{send_json_request}
\title{Send JSON POST request to the given service}
\usage{
send_json_request(json, service = "", debug = FALSE)
}
\arguments{
\item{json}{\itemize{
\item The nested list resembeling JSON that should be sent to the server
}}

\item{service}{\itemize{
\item !!WIP!! The service to send the message to
}}

\item{debug}{\itemize{
\item Default to FALSE, turns on debugging messages
}}
}
\value{
Returns the results from the query
}
\description{
Sends a POST request to the given service with a jsonlike object
}
